refactor: consolidate glass and actions modules to make them have the same shape of detached glass
