+++ src/binary-window/detached-glass/index.js
@@ -1,1 +1,3 @@
+export { DetachedGlass } from './class';
 export { default } from './module';
+export { DETACHED_GLASS_ACTIONS } from './close';
