+++ src/binary-window/glass/close.js
