+++ src/binary-window/glass/index.js
@@ -1,1 +1,2 @@
+export { Glass } from './class';
 export { default, BUILTIN_ACTIONS } from './module';
