+++ src/binary-window/glass/minimize.js
